requirements.txt
bleach==1.5.0
html5lib==0.9999999
main-roads-cloud-vision==0.0.1
Markdown==2.6.9
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
protobuf==3.4.0
six==1.10.0
tensorflow==1.3.0
tensorflow-recognizer-cli==0.0.1
tensorflow-tensorboard==0.1.4
Werkzeug==0.12.2
